+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   matches the lintian hint `wrong-section-according-to-package-name`.
 - Add `Build-Profiles` as a known field.
 - Add inspection for detecting misspelled field names.
+- Add inspection for detecting missing `${shlibs:Depends}` dependency.
 
 ### Fixed
 - Typo of the archive section `electronics`, which caused the plugin to mark the correct spelling as an error and